+++ src/modules/proto_ldap_sync/sync.c
@@ -887,7 +887,7 @@ static int _sync_state_free(sync_state_t *sync)
  * @param[in] two second sync to compare.
  * @return CMP(one, two)
  */
-static int _sync_cmp(void const *one, void const *two)
+static int8_t _sync_cmp(void const *one, void const *two)
 {
 	sync_state_t const *a = one, *b = two;
